ecourseapisv1/requirement.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
cloudinary==1.44.1
cryptography==46.0.5
Django==6.0.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
pycparser==3.0
PyMySQL==1.1.2
pytz==2026.1.post1
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
